// src/generated/models/DCPO_LISTE_DIRECTIONModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface DCPO_LISTE_DIRECTIONBase {
  ID?: number;
  Title?: string;
  libelle?: string;
  sigle?: string;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL vers une miniature présentant une version fortement compressée de l’élément, si cela est possible.
    Small?: string;
    // URL vers la miniature de taille standard de l’élément, le cas échéant.
    Medium?: string;
    // URL vers la miniature la plus grande de l’élément, le cas échéant.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface DCPO_LISTE_DIRECTIONRead extends DCPO_LISTE_DIRECTIONBase {
  Author?: AuthorValue;
  Editor?: EditorValue;
}

export interface DCPO_LISTE_DIRECTIONWrite extends DCPO_LISTE_DIRECTIONBase {
  Author?: AuthorValue;
  Editor?: EditorValue;
}